refactor: split mcp_json.rs — extract file I/O into mcp_io.rs
Pure JSON logic (entry builder, inject, initial content) stays in
mcp_json.rs; async file read/write lives in new mcp_io.rs. Each layer
now has its own focused tests. Closes #397.
