+++ src/server/handlers/init/mcp_json.rs
@@ -1,54 +1,40 @@
 use serde_json::{json, Value};
-use std::path::Path;
-use tokio::fs;
 
-const MCP_JSON_FILENAME: &str = ".mcp.json";
+pub const MCP_JSON_FILENAME: &str = ".mcp.json";
 
-fn centy_mcp_entry() -> Value {
+pub fn centy_mcp_entry() -> Value {
     json!({
         "command": "npx",
         "args": ["-y", "centy-mcp"]
     })
 }
 
-/// Ensures `.mcp.json` in the project root contains the centy MCP server entry.
+/// Build the initial `.mcp.json` content with only the centy entry.
+pub fn initial_mcp_json() -> String {
+    let content = json!({
+        "mcpServers": {
+            "centy": centy_mcp_entry()
+        }
+    });
+    let mut formatted = serde_json::to_string_pretty(&content)
+        .unwrap_or_default();
+    formatted.push('\n');
+    formatted
+}
+
+/// Inject the centy MCP entry into the parsed JSON if absent.
 ///
-/// Behavior:
-/// - File does not exist → create it with the centy entry
-/// - File exists, `mcpServers.centy` absent → inject only the `centy` key
-/// - File exists, `mcpServers.centy` already present → no-op
-/// - File exists but invalid JSON → return an error, do not modify
-pub async fn ensure_mcp_json(project_path: &Path) -> Result<(), String> {
-    let mcp_path = project_path.join(MCP_JSON_FILENAME);
-
-    if !mcp_path.exists() {
-        let content = json!({
-            "mcpServers": {
-                "centy": centy_mcp_entry()
-            }
-        });
-        let mut formatted = serde_json::to_string_pretty(&content)
-            .map_err(|e| format!("Failed to serialize .mcp.json: {e}"))?;
-        formatted.push('\n');
-        fs::write(&mcp_path, formatted)
-            .await
-            .map_err(|e| format!("Failed to write .mcp.json: {e}"))?;
-        return Ok(());
-    }
-
-    let raw = fs::read_to_string(&mcp_path)
-        .await
-        .map_err(|e| format!("Failed to read .mcp.json: {e}"))?;
-
+/// - Returns `Ok(None)` if `mcpServers.centy` is already present (no-op).
+/// - Returns `Ok(Some(updated))` with the formatted JSON string when the entry was injected.
+/// - Returns `Err` if `raw` is invalid JSON or the root is not a JSON object.
+pub fn inject_centy_entry(raw: &str) -> Result<Option<String>, String> {
     let mut doc: Value =
-        serde_json::from_str(&raw).map_err(|e| format!(".mcp.json contains invalid JSON: {e}"))?;
+        serde_json::from_str(raw).map_err(|e| format!(".mcp.json contains invalid JSON: {e}"))?;
 
-    // If mcpServers.centy already present → no-op
     if doc.get("mcpServers").and_then(|s| s.get("centy")).is_some() {
-        return Ok(());
+        return Ok(None);
     }
 
-    // Inject the centy key, preserving all other content
     let root = doc
         .as_object_mut()
         .ok_or_else(|| ".mcp.json root is not a JSON object".to_string())?;
@@ -75,89 +61,51 @@ pub async fn ensure_mcp_json(project_path: &Path) -> Result<(), String> {
     let mut formatted = serde_json::to_string_pretty(&doc)
         .map_err(|e| format!("Failed to serialize updated .mcp.json: {e}"))?;
     formatted.push('\n');
-    fs::write(&mcp_path, formatted)
-        .await
-        .map_err(|e| format!("Failed to write .mcp.json: {e}"))?;
-
-    Ok(())
+    Ok(Some(formatted))
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use tempfile::tempdir;
 
-    #[tokio::test]
-    async fn creates_mcp_json_when_absent() {
-        let dir = tempdir().unwrap();
-        ensure_mcp_json(dir.path()).await.unwrap();
-        let path = dir.path().join(".mcp.json");
-        assert!(path.exists());
-        let content = tokio::fs::read_to_string(&path).await.unwrap();
+    #[test]
+    fn initial_mcp_json_has_centy_entry() {
+        let content = initial_mcp_json();
         let doc: Value = serde_json::from_str(&content).unwrap();
         assert_eq!(doc["mcpServers"]["centy"]["command"], "npx");
         assert_eq!(doc["mcpServers"]["centy"]["args"][0], "-y");
         assert_eq!(doc["mcpServers"]["centy"]["args"][1], "centy-mcp");
     }
 
-    #[tokio::test]
-    async fn injects_centy_into_existing_file_without_centy() {
-        let dir = tempdir().unwrap();
-        let path = dir.path().join(".mcp.json");
-        let existing = r#"{"mcpServers":{"other":{"command":"other-cmd","args":[]}}}"#;
-        tokio::fs::write(&path, existing).await.unwrap();
-
-        ensure_mcp_json(dir.path()).await.unwrap();
-
-        let content = tokio::fs::read_to_string(&path).await.unwrap();
-        let doc: Value = serde_json::from_str(&content).unwrap();
+    #[test]
+    fn inject_adds_centy_to_existing_servers() {
+        let raw = r#"{"mcpServers":{"other":{"command":"other-cmd","args":[]}}}"#;
+        let updated = inject_centy_entry(raw).unwrap().expect("should inject");
+        let doc: Value = serde_json::from_str(&updated).unwrap();
         assert_eq!(doc["mcpServers"]["centy"]["command"], "npx");
-        // Other keys preserved
         assert_eq!(doc["mcpServers"]["other"]["command"], "other-cmd");
     }
 
-    #[tokio::test]
-    async fn no_op_when_centy_already_present() {
-        let dir = tempdir().unwrap();
-        let path = dir.path().join(".mcp.json");
-        let existing = r#"{"mcpServers":{"centy":{"command":"custom","args":["--custom"]}}}"#;
-        tokio::fs::write(&path, existing).await.unwrap();
-
-        ensure_mcp_json(dir.path()).await.unwrap();
-
-        let content = tokio::fs::read_to_string(&path).await.unwrap();
-        let doc: Value = serde_json::from_str(&content).unwrap();
-        // Existing custom config must not be overwritten
-        assert_eq!(doc["mcpServers"]["centy"]["command"], "custom");
-        assert_eq!(doc["mcpServers"]["centy"]["args"][0], "--custom");
+    #[test]
+    fn inject_returns_none_when_centy_present() {
+        let raw = r#"{"mcpServers":{"centy":{"command":"custom","args":["--custom"]}}}"#;
+        let result = inject_centy_entry(raw).unwrap();
+        assert!(result.is_none());
     }
 
-    #[tokio::test]
-    async fn errors_on_invalid_json() {
-        let dir = tempdir().unwrap();
-        let path = dir.path().join(".mcp.json");
-        tokio::fs::write(&path, b"not valid json").await.unwrap();
-
-        let result = ensure_mcp_json(dir.path()).await;
+    #[test]
+    fn inject_errors_on_invalid_json() {
+        let result = inject_centy_entry("not valid json");
         assert!(result.is_err());
-        let err = result.unwrap_err();
-        assert!(err.contains("invalid JSON"), "error was: {err}");
-        // File must not be modified
-        let still_invalid = tokio::fs::read_to_string(&path).await.unwrap();
-        assert_eq!(still_invalid, "not valid json");
+        assert!(result.unwrap_err().contains("invalid JSON"));
     }
 
-    #[tokio::test]
-    async fn idempotent_when_called_twice() {
-        let dir = tempdir().unwrap();
-        ensure_mcp_json(dir.path()).await.unwrap();
-        let first = tokio::fs::read_to_string(dir.path().join(".mcp.json"))
-            .await
-            .unwrap();
-        ensure_mcp_json(dir.path()).await.unwrap();
-        let second = tokio::fs::read_to_string(dir.path().join(".mcp.json"))
-            .await
-            .unwrap();
-        assert_eq!(first, second);
+    #[test]
+    fn inject_adds_mcp_servers_key_when_absent() {
+        let raw = r#"{"otherKey": 42}"#;
+        let updated = inject_centy_entry(raw).unwrap().expect("should inject");
+        let doc: Value = serde_json::from_str(&updated).unwrap();
+        assert_eq!(doc["mcpServers"]["centy"]["command"], "npx");
+        assert_eq!(doc["otherKey"], 42i32);
     }
 }
