+++ src/server/handlers/init/mod.rs
@@ -1,5 +1,6 @@
 mod config_merge;
 mod handler;
+mod mcp_io;
 mod mcp_json;
 mod post_init;
 pub use handler::{init, is_initialized};
